+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `gpio`: port and pin generics first, then mode, `PinMode` for modes instead of pins, other cleanups
 
+### Breaking changes
+
+- Passing the `Clock` parameter to `Serial` by reference
+
 ## [v0.9.0] - 2022-03-02
 
 ### Added